Practice writing interpreter
I followed the tutorial by Ruslan Spivak "Let's Build a Tutorial". The tutorial is given in Python, so I have been working on converting the code to C++. This is not a completely debugged version in C++, but the Python code works. Although this is not directly relevant to the project, this helped me understand parsing, tokenizing and lexical analysis, which will be useful towards the completion of the rest of the project.