Undate of get_type() and get_subtype() for AST
I added two abstract methods get_type() and get_subtype() in AST class.

Get_type() is defined in Statement and Expression and inherited by every single class in this file.
If, for example, you call this method on the object of the class While, it will return “Statement”.

Get_subtype() is defined in subclasses inherited from Statement and Expression such as Block, ... .
And if for example you call this method on the object of the class While, it will return “While”.